Renames helpers module to builtins
